dynamic pricing strategy urgencypricing

+++ src/main/java/org/cloudspiretech/in/LunaCordis/strategy/PricingStrategy.java
@@ -2,6 +2,8 @@ package org.cloudspiretech.in.LunaCordis.strategy;
 
 import org.cloudspiretech.in.LunaCordis.entity.Inventory;
 
+import java.math.BigDecimal;
+
 public interface PricingStrategy {
-    public double calculatePrice(Inventory inventory);
+    public BigDecimal calculatePrice(Inventory inventory);
 }
